[ADD] Board - 파라미터 입력 값 검증 추가

+++ controllers/boardController.js
@@ -2,6 +2,7 @@
 
 const boardService = require('../services/boardService');
 const { parameterPresenceCheck } = require('../common/modules');
+const {ParameterError} = require("../common/modules/customErrors");
 
 const checkArray = {
     createAPost : ['content', 'password', 'isPrivate'],
@@ -22,7 +23,8 @@ exports.addPost = async (req, res, next) => {
 
 exports.findAllPost = async (req, res, next) => {
     try  {
-        const { offset, limit } = req.body;
+        const { offset, limit } = req.query;
+        if(parseInt(offset)<0 || parseInt(limit)<0) throw new ParameterError;
         parameterPresenceCheck(checkArray.findAllPosts, {offset, limit});
         const posts = await boardService.findAllPost(parseInt(offset), parseInt(limit));
         return next({status: 'success', posts: posts});
@@ -34,6 +36,7 @@ exports.findAllPost = async (req, res, next) => {
 exports.findOnePost = async (req, res, next) => {
     try  {
         const postId = req.params.id;
+        if(parseInt(postId)<0) throw new ParameterError;
         const post = await boardService.findOnePost(postId);
         return next({status: 'success', posts: post});
     } catch (e) {
@@ -44,6 +47,7 @@ exports.findOnePost = async (req, res, next) => {
 exports.editPost = async (req, res, next) => {
     try  {
         const postId = req.params.id;
+        if(parseInt(postId)<0) throw new ParameterError;
         const { content, password } = req.body;
         parameterPresenceCheck(checkArray.editPost, {postId, content, password});
         const post = await boardService.editPost(req.userId, postId, content, password);
@@ -56,6 +60,7 @@ exports.editPost = async (req, res, next) => {
 exports.deletePost = async (req, res, next) => {
     try  {
         const postId = req.params.id;
+        if(parseInt(postId)<0) throw new ParameterError;
         await boardService.deletePost(req.userId, postId);
         return next({status: 'success'});
     } catch (e) {
